refactor: Import spammer and storageprefix with new path
